kirin970: initial import from CLT-AL00-AL01-AL00l-TL00-TL01_02_HM_opensource.tar.gz
Changes:
* find . -type f -exec dos2unix {} \;

+++ tools/perf/util/trace-event-info.c
@@ -379,7 +379,7 @@ out:
 
 static int record_saved_cmdline(void)
 {
-	unsigned long long size;
+	unsigned int size;
 	char *path;
 	struct stat st;
 	int ret, err = 0;
